Update frontend: add webpack
