+++ README.md
@@ -6,6 +6,7 @@ Backend на Go - Igor Efremov
 
 1. go build -o main
 2. ./main
-3. browserify static/lib/main.js -o static/bundle.js
-4. [`http://localhost:3000/`](http://localhost:3000/)
-
+3. cd web
+4. npm install
+5. npm run build
+6. [`http://localhost:3000/`](http://localhost:3000/)
