--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "try-to-game",
-  "version": "1.0.0",
-  "description": "Backend на Go - Igor Efremov",
-  "main": "static/lib/main.js",
-  "directories": {
-    "lib": "lib"
-  },
-  "dependencies": {
-    "pixi.js": "^5.2.4"
-  },
-  "devDependencies": {
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11"
-  },
-  "scripts": {
-    "build": "webpack"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
