+++ static/index.html
@@ -1,52 +1,60 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<title>Game</title>
-	<link rel="stylesheet" href="/static/main.css">
-</head>
-<body>
+  <head>
+    <meta charset="UTF-8" />
+    <title>Game</title>
+    <link rel="stylesheet" href="/static/css/main.css" />
+  </head>
+  <body>
+    <div hidden id="greet" class="view-element greet-view">
+      <div class="container">
+        <form class="greet-form">
+          <label for="username">Представьтесь</label>
+          <input
+            type="text"
+            class="greet-form__username-input"
+            id="username"
+            name="username"
+            placeholder="Введите ваше имя"
+          />
+          <div class="greet-form__choose-buttons">
+            <input
+              type="button"
+              class="greet-form__choose-multi-button"
+              id="choose-multi"
+              value="Играть с другом"
+            />
+          </div>
+        </form>
+      </div>
+    </div>
 
-<div hidden id="greet" class="view-element greet-view">
-	<div class="container">
-		<form class="greet-form">
-			<label for="username">Представьтесь</label>
-			<input type="text" class="greet-form__username-input" id="username" name="username"
-				   placeholder="Введите ваше имя"/>
-			<div class="greet-form__choose-buttons">
-				<input type="button" class="greet-form__choose-multi-button" id="choose-multi" value="Играть с другом">
-			</div>
-		</form>
-	</div>
-</div>
+    <div hidden id="wait" class="view-element wait-view">
+      <div class="container">
+        <div style="transform: scale(0.75);">
+          <div class="cssload-thecube">
+            <div class="cssload-cube cssload-c1"></div>
+            <div class="cssload-cube cssload-c2"></div>
+            <div class="cssload-cube cssload-c4"></div>
+            <div class="cssload-cube cssload-c3"></div>
+          </div>
+        </div>
+        <h1 class="wait-view__header">Мы ищем вам компаньона</h1>
+      </div>
+    </div>
 
-<div hidden id="wait" class="view-element wait-view">
-	<div class="container">
-		<div style="transform:scale(0.75);">
-			<div class="cssload-thecube">
-				<div class="cssload-cube cssload-c1"></div>
-				<div class="cssload-cube cssload-c2"></div>
-				<div class="cssload-cube cssload-c4"></div>
-				<div class="cssload-cube cssload-c3"></div>
-			</div>
-		</div>
-		<h1 class="wait-view__header">Мы ищем вам компаньона</h1>
-	</div>
-</div>
+    <div hidden id="app" class="view-element game-view"></div>
 
-<div hidden id="app" class="view-element game-view">
-</div>
-
-<div hidden id="finish" class="view-element finish-view">
-	<div class="pyro">
-		<div class="before"></div>
-		<div class="after"></div>
-	</div>
-	<div class="finish-view__results">
-		<h1 class="finish-view__header">Игра окончена</h1>
-		<h2 class="finish-view__result"></h2>
-	</div>
-</div>
-<script src="/static/bundle.js"></script>
-</body>
+    <div hidden id="finish" class="view-element finish-view">
+      <div class="pyro">
+        <div class="before"></div>
+        <div class="after"></div>
+      </div>
+      <div class="finish-view__results">
+        <h1 class="finish-view__header">Игра окончена</h1>
+        <h2 class="finish-view__result"></h2>
+      </div>
+    </div>
+    <script src="/static/js/bundle.js"></script>
+  </body>
 </html>
