+++ web/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "try-to-game",
+  "version": "1.0.0",
+  "description": "Backend на Go - Igor Efremov",
+  "main": "src/main.js",
+  "dependencies": {
+    "browserify": "^16.5.1",
+    "pixi.js": "^5.3.0"
+  },
+  "scripts": {
+    "build": "./node_modules/.bin/webpack",
+    "watch": "./node_modules/.bin/webpack -w",
+    "server": "./node_modules/.bin/webpack-dev-server"
+
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
+  }
+}
