+++ web/webpack.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  mode: "development",
+  entry: "./src/index.js",
+  output: {
+    filename: "../../static/js/bundle.js",
+  },
+  devServer: {
+    contentBase: "../",
+  },
+  devtool: 'inline-source-map'
+};
